Remove POST /oauth/authorize endpoint
I'm fairly confident this hasn't been used since bc9c1f017 (14 Jan 2013)
when we configured Doorkeeper to skip authorization in the `new` action
of this controller. Because `skip_authorization?` always returned `true`
we were never rendering the `new.html.erb` template which meant that
there was no way for an end user to POST to the `#create` action by
using the user interface. While it was technically possible for someone
to use another tool (e.g. curl) to POST to this endpoint that's not a
use case we need to support.

I can't see any POST requests to /oauth/authorize in Logit. Although
Logit only contains data for the past week or two I'd still expect to
see requests if this endpoint was in use.

Ideally I'd remove this route entirely but that doesn't appear to be
possible with Doorkeeper's configuration options so I'm explicitly
returning a 404 instead.