Remove DELETE /oauth/authorize endpoint
I can't see any POST requests to /oauth/authorize in Logit. Although
Logit only contains data for the past week or two I'd still expect to
see requests if this endpoint was in use.

Ideally I'd remove this route entirely but that doesn't appear to be
possible with Doorkeeper's configuration options so I'm explicitly
returning a 404 instead.