Migrate UsersController#require_2sv page to design system
I've used the standard hint, checkbox and button components.

This page is shown as part of the user creation journey for users
where their organisation doesn't mandate 2sv but we want to mandate it
for the specific user. I've decided not to include a breadcrumb trail
as there's not really any consideration made for navigating forward
and backwards in this journey (the record is saved and the user
emailed before we reach this page, for example).