Replace <del> tag with explanatory text
We want to replace this table with the the design system table
component[1] but that doesn't support HTML tags in table cells. This
commit replaces the <del> tag with some explanatory text in
parenthesis. I decided it was better to have different text for each
of the three cases rather than something like "(unusable account)" so
have inlined the three cases provided by `User.unusable_account?`
method into this helper.

To retain the visual indication provided by <del> I've added a custom
class, this will make it easier to switch to using a cell "format"
when we move to the design system. Note that CSS `line-through` is
not, as far as I can tell, accessible to screenreaders so it's useful
to have the explanatory text in parenthesis as well as the visual
indication.

[1] https://components.publishing.service.gov.uk/component-guide/table