Move 2SV exemptions page to use GOV.UK Design System
Trello: https://trello.com/c/oqt4y7dH

I've introduced a new TwoStepVerificationExemption ActiveModel model to
implement the field validation for this form; previously the validation
was all done in the controller.

My main motivation for doing this was to be able to store validation
errors against a particular field so that I could implement the Design
System error summary correctly with links to the invalid fields and so I
could implement something close to the date field validation recommended
by the Design System [1].

However, I think it also bring some other benefits: the code is now a
bit more idiomatic Rails, it's easier to test the validation, and when
validation fails the values the user entered are retained (i.e. we
re-render the form rather than redirecting as we did previously).

I've intentionally implemented the expiry_date validation in a
"heirarchical" manner, i.e. the user doesn't immediately see a bunch of
validation errors as they would if I'd used standard Rails validation.
Instead they see the most "important" validation error(s) first and then
if they fix those, they might see less "important" validation errors. I
haven't actually seen any guidance on this, but it felt overwhelming for
the user to see many different validation errors for a single date field
when e.g. the problem is that none of the values have been entered at
all.

The Design System documentation re validation errors on a date input [2]
recommends highlighting individual inputs within the date input (e.g.
just the day field) if it's possible to determine that's where the
problem lies. Although the new TwoStepVerificationExemption model makes
it feasible to do this, the date_input component [3] in the
govuk_publishing_components gem doesn't currently support setting an
error on the individual day, month & year fields. I plan to suggest
changing the date_input component to support this in the future.

I've haven't included a Cancel button, because we now have a decent
breadcrumb trail. This is what we've done elsewhere when converting
pages to use the GOV.UK Design System.

[1]: https://design-system.service.gov.uk/components/date-input/
[2]: https://design-system.service.gov.uk/components/date-input/#error-messages
[3]: https://github.com/alphagov/govuk_publishing_components/blob/45ea966343920fb73990a4c3b3ed517b76219d56/app/views/govuk_publishing_components/components/_date_input.html.erb