Move users_with_access page to design system
This commit converts
DoorkeeperApplicationsController#users_with_access to the design
system.

I've decided to use a single pagination component as that seems to be
more in keeping with the relevant pattern[1].

The `format` key on each table cell converts the passed in string to a
CSS class name (e.g. `email` becomes `.govuk-table__cell--email`), and
adds no class when nil is provided, so I was able to simplify
formatted_user_name_class. I intend to rename that method in the next
commit to better suit its new purpose.

[1] https://design-system.service.gov.uk/components/pagination/