Move application permissions index page to use Design System
Trello: https://trello.com/c/S2Gq9Nrn

I've moved the database query to obtain the list of supported
permissions into a controller action, because I think that's a bit more
idiomatic.

I haven't linkified the name of each permission as it was before,
because the link is the same as for the "Edit" action on the same row
which seemed like overkill and made the table harder to scan.

I've had to tweak the assertions in the controller test, because it's
not straightforward to add the same CSS classes to the table cells.
Since these classes were only added for the assertion selectors, I don't
think this is a bad thing.