Remove "Sign-in count" column from users page
Trello: https://trello.com/c/QuwuPtZW

We're trying to simplify/improve this page a bit prior to changing it to
use the GOV.UK Design System. We don't think this column is very useful
and it's cluttering up the table unnecessarily. We did contemplate
adding the information to the individual (edit) user page, but it didn't
seem useful enough and in any case you can probably find out what you
want to know from the "Account access log" page for the user.
Furthermore, this change won't remove the "Sign-in count" column from
the CSV file generated by clicking the "Export as CSV" button.