Remove "Created" column from users page
Trello: https://trello.com/c/QuwuPtZW

We're trying to simplify/improve this page a bit prior to changing it to
use the GOV.UK Design System. We don't think this column is very useful
and it's cluttering up the table unnecessarily. We did contemplate
adding the information to the individual (edit) user page, but it didn't
seem useful enough and in any case this change won't remove the
"Created" column from the CSV file generated by clicking the "Export as
CSV" button.