Remove "2SV Exemption Expiry" column from users page
Trello: https://trello.com/c/QuwuPtZW

We're trying to simplify/improve this page a bit prior to changing it to
use the GOV.UK Design System. This information is available on the
individual (edit) user page anyway, so we don't think this column is
useful enough to justify retaining it on this page.

Note that this change won't remove the "2SV Exemption Expiry Date"
column from the CSV file generated by clicking the "Export as CSV"
button.