Always use standard pagination on users page
Trello: https://trello.com/c/QuwuPtZW

We don't think the "User by initial" alphabetical pagination which was
shown when no filters were selected was very useful. Also it separated
the filters from the table which means that it was less obvious that the
filters applied to the table.

Previously the alphabetical pagination ordered the users alphabetically
by name when no filters were selected, but as soon as a filter was
selected the users were in an undefined (natural database) order. Now
the users are always ordered alphabetically by name whether or not any
filters are selected.

Note that I've had to make a bunch of changes to
test/integration/admin_user_index_test.rb, so that the assertions match
how the users created in the test setup is now presented in the page.