Remove pagination controls from top of users page
Trello: https://trello.com/c/QuwuPtZW

And reduce number of users per page from 100 to 25.

We're trying to move the filter controls closer to the table which they
affect. Removing the pagination controls from the top of the table is a
step in this direction. Reducing the number of users per page to 25
makes the pagination controls at the bottom of the page more obvious.