Move "Export as CSV" outside filter on users page
Trello: https://trello.com/c/QuwuPtZW

Previously it was confusing that the button was "inside" the filter
controls and it was also quite easy to click the button when you meant
to click the "X" to remove all filters.

Having the button level with the sub-heading which includes the number
of users gives a visual clue that the CSV will also contain that number
of users, i.e. the filters apply to the CSV file as well as the table in
the page.