Split name & email into separate columns on users page
Trello: https://trello.com/c/QuwuPtZW

We're planning to move this page to use the GOV.UK Design System and the
table component for the Design System doesn't work well with line break
elements inside a cell. We've recently removed some columns from the
table and so there's a bit more room to have these as two columns now.