Use Mocha vs Minitest Mock in inviting users test
The use of MiniTest::Mock in this test was causing a problem when trying
to upgrade to Minitest v5.19.0, because we should be using the newer
Minitest module name.

I could have just changed that here, but it seemed odd that we're using
Minitest mocking & stubbing in this one test, while in every other test
we're using Mocha. So I decided to convert this test to use Mocha too.

I'm pretty confident that we don't need to "expect" the calls to the
code & body methods on the response since their values are read and
asserted against in the assert_response_contains call, so using stubbed
return values is sufficient.

I think these changes also make the test easier to read.