Move name & email filter to LHS on users page
Trello: https://trello.com/c/QuwuPtZW

We think this filter is more useful that the other drop-down filters and
in user research some users are not discovering it, so we've moving it
to the left-hand side to make it more prominent.

The "Name or email" label didn't make so much sense when it was to the
left of the "Filter by" label, so we've hidden it from users (but not
from screen-readers) and added it as placeholder text for the text
input.

I've also introduced a custom color for the placeholder text to ensure
it's as contrast-y as possible (for accessibility) while still being
lighter than other text on the page. I've made the selector for the
relevant CSS rule specific to this one text input to limit the scope of
the change which isn't ideal, but I'm not concerned about that because
we're planning to move this whole app away from govuk_admin_template and
onto the GOV.UK Design System at which point the legacy CSS will be
deleted!

We've also changed the button text from "Search" to "Filter", because we
think that makes it's function clearer.