Strip unwanted whitespace from User names
We have some cases of `User` names with leading and/or trailing spaces
in our production data. Before addressing those, we'd like to prevent
any new occurrences.

Users are created in a couple of different ways and from a few different
places in this codebase. I've tried to sidestep all of that and make
a small change in a single place.