Strip unwanted whitespace from batch user names
As with regular `User`s, in our production data we also have some cases
of `BatchInvitationUser` records with leading and/or trailing whitespace
in their names. It's not as big an issue as it is with `User`s (since
a `BatchInvitationUser` isn't as central or as long-lived a concept in
the app) but I figured that we might as well fix both while we're at it.
And before addressing the production data, we'll prevent any new
occurrences.

The choice of where to do this is based on how we did it for `User`.