Remove unwanted whitespace from existing records
We have some cases of `User` and `BatchInvitationUser` records in our
production data that have leading and/or trailing whitespace in their
name columns and we'd like to remove that whitespace -- because it
"breaks" alphabetical sorting by name and it's just sorta messy.

We've just introduced `before_save` hooks to both of their respective
Rails models which will strip that unwanted whitespace anytime a record
is saved. Since there are only a small number of records that need
fixing*, it seems reasonable to load those objects and re-save them to
trigger those hooks.

* ~400 `User`, I'm not sure about `BatchInvitationUser`