Validate batch invitation users one-by-one
Until now, one invalid email address in a batch upload (CSV) would raise
an ambiguous error and give the user no information about what went
wrong.

Now, if the batch encounters a validation problem, it returns to the
original form as before but also displays a more helpful error message.

This is a small improvement which avoids some of the complexities of
what might be even more helpful error reports (e.g. listing one or all
of the invalid email addresses).