Remove Doorkeeper client_credentials grant flow
Trello: https://trello.com/c/Jx1718RH

By default Doorkeeper is configured to support both authorization_code &
client_credentials grant flows.

The authorization_code grant flow is being used e.g. when a user signs
in from a publishing application. However, it was unclear whether the
client_credentials flow was being used at all.

When a Doorkeeper::AccessToken is created as part of the client
credentials grant flow, the resource owner is always set to nil.
However, there is a NOT NULL database constraint on
oauth_access_tokens.resource_owner_id which triggers an
ActiveRecord::NotNullViolation exception.

That NOT NULL contraint has been in place since the Doorkeeper-related
tables were first added, presumably generated by a very early version of
the Doorkeeper generator [1]. However, a few minor versions of
Doorkeeper later, that same constraint was removed [2], but not removed
from Signon. This Doorkeeper release was also the same release where
the client credentials grant type was added [3].

I'm pretty confident this means that the client credentials grant flow
has never worked in Signon. Rather than fixing it, this indicates that
we can't need it and so I'm removing it in this commit to reduce the
"surface area" for any attacks.

[1]: https://github.com/doorkeeper-gem/doorkeeper/commit/fb76738381f27fe4cb62d126c1f41d45bba77dce
[2]: https://github.com/doorkeeper-gem/doorkeeper/commit/66ba66d9fecae64b9e3e8918830f682bf0df718e
[3]: https://github.com/doorkeeper-gem/doorkeeper/blob/b3d6787341561e5bc51336e3e04622ca131d6445/CHANGELOG.md#040---2012-05-26