Hide user research recruitment banner permanently
If the user clicks on "Sign up to take part in research" then we assume
that they've completed the Google Form and therefore we don't want to
annoy them by continuing to show them the recruitment banner.

I've introduced the `#user_research_recruitment_form` action so that we
can record that the user has clicked the button before sending them off
to the Google Form.