Use responsive spacing helpers for UR recruitment banner
Since we want this banner to look sensible on all screen widths, it
seems to make sense to use the GOV.UK Design System responsive spacing
helpers [1] rather than hard-coded values.

I've also replaced the use of margin-top on the divider with the
govuk-!-margin-top-0 responsive spacing override class [2], because that
seems to be the recommended way to do things.

[1]: https://design-system.service.gov.uk/styles/spacing/#responsive-spacing
[2]: https://design-system.service.gov.uk/styles/spacing/#responsive-spacing-override-classes