Improve vertical alignment of UR recruitment buttons
Based on feedback from Calum [1]:

> Also govuk-button-group could probably use some better horizontal
> alignment for the button and hide this text by using Flexbox on the
> parent element from the medium size breakpoint upward

Note that I've assumed Calum was referring to the vertical alignment
(not the horizontal alignment) of the "Sign up..." button and the "Hide
this" link. The govuk-button-group was already using a flexbox layout.
However, the align-items property was set to "baseline" and this seemed
to be why the link was appearing a bit "higher" up than the button.

Setting the align-items property to "center" seems to give better
vertical alignment.

[1]: https://trello.com/c/yGMU3sR2/211-add-banner-to-recruit-people-for-user-research#comment-64d0dadd8631bb229bbb6600