Set autocomplete attribute on input elements appropriately
Trello: https://trello.com/c/WHsgcFDz

This explicitly sets the autocomplete attribute to "off" on all input
elements unless there is a specificly suitable value, e.g. "name",
"email", "current-password", "new-password" [1]. Note that these values
must apply for the current user and not for another user, e.g. when a
user is editing the details of another user, autocomplete should be set
to "off" for the "name" & "email" fields.

The original motivation behind this was that some password managers were
filling in the "name" and "email" fields on the edit user page. This was
because autocomplete was not set on the fields, the password manager
heuristics decided that they were legitimate "name" & "email" fields,
and (since there was only one matching item for the domain) the password
manager filled the field with the current user's name and email address.
This should no longer happen, because we are now setting autocomplete to
"off" for those two fields.

[1]: https://developer.mozilla.org/en-US/docs/Web/HTML/Attributes/autocomplete#values