Add user research recruitment banner
Trello: https://trello.com/c/yGMU3sR2

We're only showing the banner on the dashboard so that we don't
overwhelm people by showing it everywhere.

We're using larger, bold text and a "start" button [1] to make the "Find
out more" CTA stand out over the "Hide this" button.

Ideally we'd have used the GOV.UK Components button [2] for the two
buttons. However, the `govuk-button--inverse` CSS class [3] is not
supported by the component and there's no easy way to override it.

Instead of using that component, we're explicitly writing out the HTML
that it would have generated. It would be worth opening a PR to add an
`inverse` property to the GOV.UK Components button so we can avoid this
duplication.

[1]: https://design-system.service.gov.uk/components/button/#start-buttons
[2]: https://components.publishing.service.gov.uk/component-guide/button
[3]: https://design-system.service.gov.uk/components/button/#buttons-on-dark-backgrounds