Add user research recruitment banner
Trello: https://trello.com/c/yGMU3sR2

We're only showing the banner on the dashboard so that we don't
overwhelm people by showing it everywhere.

Ideally we'd have used the GOV.UK Components button [1] for the two
buttons. However, the `govuk-button--inverse` CSS class [2] is not
supported by the component and there's no easy way to override it.
Since we're not using that component we've had to introduce
`ApplicationHelper#arrow_svg` to generate the arrow SVG for the
start-style button [3]. It would be worth opening a PR to add an
`inverse` property to the GOV.UK Components button so we can avoid this
duplication.

[1]: https://components.publishing.service.gov.uk/component-guide/button
[2]: https://design-system.service.gov.uk/components/button/#buttons-on-dark-backgrounds
[3]:
https://design-system.service.gov.uk/components/button/#start-buttons