Hide user research recruitment banner permanently
Trello: https://trello.com/c/yGMU3sR2

If the user clicks on "Find out more" then we assume that they've
completed the Google Form and therefore we don't want to annoy them by
continuing to show them the recruitment banner.

We've introduced the `UserResearchRecruitmentController#participate`
action so that we can record that the user has clicked the button before
sending them off to the Google Form.