Extract GovukEnvironment.name method
I want to encapsulate as much of the behaviour relating to the
INSTANCE_NAME env var as possible before I fix a bug relating to it.

I think the combination of this class name & method name better
describes its purpose and it's also a bit clearer where to find the
logic for it. As such I'm comfortable with losing the comment that was
in config/initializers/instance_name.rb.

It's also allowed me to add some tests around the untested logic that
was in config/initializers/instance_name.rb. And it allows us to stub
GovukEnvironment.name instead of Rails.application.config.instance_name
which seems like a bit of an improvement.

This change does mean we're no longer "caching" the instance_name value,
but I don't think an env var lookup is particularly slow. If it turns
out to be a problem, we can always memo-ize the value.