Extract GovukEnvironment.production? method
This makes the code a lot more intention-revealing. Previously it was
extremely unclear that the conditional logic was designed to distinguish
between the production environment and staging, integration, test &
development environments.

Note that the current implementation of the method still depends on the
incorrect assumption that the INSTANCE_NAME env var is only set in the
integration & staging (not production) environments. However, I plan to
address that in a later commit.