Revoke all access tokens when suspending a user
Until now a suspended ApiUser is flagged as suspended in the database,
which results in Signon telling applications that the user has no
permissions. Here we add the extra step of revoking the user's access
tokens, too.

If tokens aren't revoked, a suspended user technically retains its
authorised access (potentially for years, because TTL is quite long).
Despite the user also losing all of their permissions, this does present
a security concern.

The intended target of this change is `ApiUser`s (i.e. applications and
not human users). The implementation is simpler if we just perform this
step for suspended `User`s more generally and it's safe to do that
because in the case of normal, human users, new tokens are created
automatically on signin if expired or revoked.