Make MailHelper#app_name consistent with #email_from_address
In MailHelper#email_from_address and other places we do NOT include the
word "production" to distinguish which GOV.UK environment (integration,
staging, production) the app is running in.

So it seems more consistent and less confusing to the user NOT to
include the word "production" in MailHelper#app_name.