Extract rake task logic to its own class
Because Rake isn't as nice to work with or test as regular objects are.

I've kept the new code as close to as-is as I think makes sense. The new
class is specific to a single OAuth access record type per-instance
purely because the rake task wants to print the total number of deleted
records for each type. I'm not precious about that reporting to STDOUT,
I just have no reason to change it