Enable OAuth2 PKCE extension in Doorkeeper
I ran `bundle exec rails generate doorkeeper:pkce` to enable the PKCE
extension[1] in Doorkeeper[2]. Rubocop complained about the migration that
was generated so I changed it to use `change_table` instead of two calls
to `add_column`.

The PKCE extension isn't mandatory at this point (as demonstrated by the
"oauth2 authorization code grant type" test continuing to pass) but it
will be used if the client supplies a `code_challenge_method` and
`code_challenge` when requesting the Authorization Code.

It's our intention to mandate the use of PKCE once we can be sure all
clients of Signon are using it.

The Doorkeeper docs[2] include the following 2 points to consider:

1. If you overrode the doorkeeper/authorizations/new.html.erb view, make
   sure that you have the code_challenge and code_challenge_method hidden
   form fields.

We never render this template and haven't overridden it so this is of no
concern to us.

2. Also, in case your client is public (e.g. mobile app, single page
   app) note that the Doorkeeper::Application for the client should have
   confidential: false.

We only support confidential clients at the moment so this is of no
concern to us. Note that we set `Doorkeeper::Application#confidential`
to true in schema.rb[3] and don't allow it to be overridden in the UI.

[1]: https://datatracker.ietf.org/doc/html/rfc7636
[2]: https://doorkeeper.gitbook.io/guides/ruby-on-rails/pkce-flow
[3]: https://github.com/alphagov/signon/blob/78ebf57fe0a787d1e462b39f2475d3b74f085511/db/schema.rb#L114