Use GOVUK_ENVIRONMENT_NAME vs INSTANCE_NAME env var
This follows on from this PR [1].

The GOVUK_ENVIRONMENT_NAME env var is made available to all apps [2] and
is already being used in this app, e.g. to configure GovukAdminTemplate.

The GOVUK_ENVIRONMENT_NAME env var has the same values as the
INSTANCE_NAME env var in the different environments, i.e.

* "integration" [3]
* "staging" [4]
* "production" [5]

So using GOVUK_ENVIRONMENT_NAME shouldn't change the behaviour of the
app in any way.

We should look to remove the INSTANCE_NAME variable from
govuk-helm-charts once this change has been deployed, because it will no
longer be needed by this app.

[1]: https://github.com/alphagov/signon/pull/2311
[2]: https://github.com/alphagov/govuk-helm-charts/blob/f117fcbc4bf103431f73a8d66b6dda086939e1b1/charts/app-config/templates/env-configmap.yaml#L22
[3]: https://github.com/alphagov/govuk-helm-charts/blob/f117fcbc4bf103431f73a8d66b6dda086939e1b1/charts/app-config/values-integration.yaml#L4
[4]: https://github.com/alphagov/govuk-helm-charts/blob/f117fcbc4bf103431f73a8d66b6dda086939e1b1/charts/app-config/values-staging.yaml#L4
[5]: https://github.com/alphagov/govuk-helm-charts/blob/f117fcbc4bf103431f73a8d66b6dda086939e1b1/charts/app-config/values-production.yaml#L4