Add test to exercise OAuth2 flow
This test demonstrates/exercises the OAuth2 interactions involved in the
Authorization Code Grant type[1]. This is the "happy path" that's
followed when the user has the signin permission in the application
they're authorizing to make requests on their behalf.

I'm adding this to give me confidence that I'm not breaking this
behaviour when I enable the PKCE extension.

Note that some of our "integration" tests are using Rails'
RequestHelpers[2] (e.g `get` and `post`) while others are using the
Capybara methods (e.g. `visit`). I'm using the Rails RequestHelpers as I
need to simulate requests that happen outside of the browser (e.g.  the
request for the Access Token).

Note that CSRF protection was enabled in our integration tests in
c69288d1516714ba2c4fe6ad30e0f4f18bc17bf7. It's not particularly useful
in these tests so I'm disabling it rather than working out how to
get/set the authenticity token.

I'm using `ActionDispatch::Integration::Runner#reset!`[3] to reset the
session between the user approving the Authorization Grant (in the form
of the Authorization Code) and the client using that Authorization Grant
to request an Access Token. My hope is that this makes it slightly
clearer that the Authorization Code request is performed by the user's
User Agent and the Access Token request, and subsequent request for the
protected resource, is performed by the client.

[1]: https://oauth.net/2/grant-types/authorization-code/
[2]: https://api.rubyonrails.org/v7.0.7/classes/ActionDispatch/Integration/RequestHelpers.html
[3]: https://api.rubyonrails.org/v7.0.7/classes/ActionDispatch/Integration/Runner.html#method-i-reset-21