Publish publicly accessible /privacy-notice page
To meet one of the requirements of our Data Protection Impact
Assessment, we're publishing this privacy notice and cookie policy for
users of Signon.

I've specified a distinct page title and page heading to avoid the
duplication of the app's name in the <title>.

I've used as many GOV.UK publishing components in the building of this
page as made sense to me.

To be honest, I have very little understanding of the differences
between public and admin styles and no understanding of the thinking
behind those differences. For those reasons, I'm not 100% confident how
this page should look.

The page uses the publishing components Admin Layout (as opposed to the
Public Layout), because all of Signon uses it and this page isn't public
facing even though it takes some cues from public facing pages (e.g.
https://www.gov.uk/help/privacy-notice).

Because the Admin Layout doesn't appear to style links the way that we
expect them to look on GOV.UK pages (with content similar to this one)
and because it doesn't appear to enforce the use of the GDS Transport
font family, I've used `govuk-body` and `govuk-link` CSS classes to
achieve the same effect.

For the published date line at the end of the page, I've styled the text
to match how that information looks on public facing GOV.UK pages. There
is a Published Date component that lives in the Government Frontend
codebase (not shared through the components gem), but the effort to move
that into the gem only to not end up using any of its features (for now)
didn't seem worth it.